--------------------------- main.cc --------------------------- Adam added a featured to main.cc that logs the location and direction that the camera is pointing in. This information is written to a file so that it can be refrenced at a later point.
Adam also added functionality to main.cc to allow it to generate a path for the camera to follow automatically. This function works by using the distance map implemented by Taylor in renderer.cc (more to come on this later). It works pretty well... it will give us a good idea of where to go.

---------------------------
renderer.cc
---------------------------
Taylor add fucntionality that generates a depth map. This allows the path planning for the camera to make use of depth information gathered from the raymarcher. In the future this information will also be used to increase the realness of the image by adding fogg and nice shadows (hopefully)

Funcationality for debugging where the algorithm checks for the max depth was also added. Details of this were documented in renderer.h.

---------------------------
color.h
---------------------------
Taylor added a distance field to the pixelData struct. This information is used in the distance map. It will become a vital piece of information for shadding and colouring.

---------------------------
getcolor.cc
---------------------------
had to modify some fucntion defs to allow for a distance to be generated per pixel. Also introduced some basic fogging but it needs a lot of work (and doesn't really work right now)

---------------------------
getparams.c
---------------------------
Had to modify this file so that it reflected all of the params that have been added to the params file. (These are all for camera path planning stuff)

---------------------------
print.c
---------------------------
Not sure what changed in this file.....

---------------------------
raymarching.cc
---------------------------
Updated so that it stores the depth (distance from view plane) of the current pixel.

---------------------------
renderer.h
---------------------------
Had to be updated to reflect all of the params being read in from the params.dat file -> needed a way to pass them to the code in renderer.cc, this seemed like the most appropriate way

---------------------------
savebmp.c
---------------------------
Not sure what changed here... I don't think anything