Added Basic Fogging
Added fogging functionality. It is based off of the maximum distance in the camera's view frame (not the entire picture). All the code does is mix in some of the background colour the closer you get to the maximum distance.

Changed main around a little bit and made the camera move towards a point correctly (have to use dot product). I left the other code (to move the camera around) in comments in case if we want to refrence it.

Next steps:
Need to get the camera to centre on a new target. It is still not doing this correctly and I am not sure why...
Work on lighting stuff. Maybe start on some shadows